Refactor common code to abstract Sensors class
